+++ WoWsShipBuilder.UI/Extensions/IScalableViewModel.cs
@@ -1,0 +1,15 @@
+﻿using Avalonia;
+using Avalonia.Platform;
+
+namespace WoWsShipBuilder.UI.Extensions
+{
+    internal interface IScalableViewModel
+    {
+        double ContentScaling { get; set; }
+    }
+
+    public interface IScalableWindow
+    {
+        void ProcessResizing(Size newSize, PlatformResizeReason resizeReason);
+    }
+}