+++ ReleaseNotes.md
@@ -5,7 +5,7 @@
 - Add optional telemetry data in settings. Telemetry data is disabled by default and needs to be enabled manually.
 
 ## Changes
-
+- Windows now automatically resize to fit on smaller screens or on screens with dpi scaling enabled (#56)
 
 ## Bugfixes
 - Show missing modules on aircraft carriers (#38)